Added vuex module state for grv
