+++ frontend/src/store/index.js
@@ -3,6 +3,7 @@ import Vuex from 'vuex'
 import general from './modules/general'
 import nodes from './modules/nodes'
 import types from './modules/types'
+import grv from './modules/grv'
 
 Vue.use(Vuex)
 
@@ -12,7 +13,8 @@ export default new Vuex.Store({
   modules: {
     general,
     nodes,
-    types
+    types,
+    grv
   },
   strict: debug,
   plugins: debug ? [] : []
