+++ frontend/src/store/modules/types.js
@@ -2,7 +2,6 @@
 const state = {
 }
 
-// getters
 const getters = {
 }
